+++ src/api/index.ts
@@ -11,35 +11,25 @@ import { patienceDiff, patienceDiffPlus } from '@/patience_diff';
 import { CONFIG, CONNECTION } from '@/config';
 
 class Config implements NeuroPilotConfig {
-    get beforeContext(): ConfigValue<number>
-    { return {settingID: 'beforeContext', value: CONFIG.beforeContext}; }
+    get beforeContext(): ConfigValue<number> { return { settingID: 'beforeContext', value: CONFIG.beforeContext }; }
 
-    get afterContext(): ConfigValue<number>
-    { return {settingID: 'afterContext', value: CONFIG.afterContext}; }
+    get afterContext(): ConfigValue<number> { return { settingID: 'afterContext', value: CONFIG.afterContext }; }
 
-    get cursorFollowsNeuro(): ConfigValue<boolean>
-    { return {settingID: 'cursorFollowsNeuro', value: CONFIG.cursorFollowsNeuro}; }
+    get cursorFollowsNeuro(): ConfigValue<boolean> { return { settingID: 'cursorFollowsNeuro', value: CONFIG.cursorFollowsNeuro }; }
 
-    get sendContentsOnFileChange(): ConfigValue<boolean>
-    { return {settingID: 'sendContentsOnFileChange', value: CONFIG.sendContentsOnFileChange}; }
+    get sendContentsOnFileChange(): ConfigValue<boolean> { return { settingID: 'sendContentsOnFileChange', value: CONFIG.sendContentsOnFileChange }; }
 
-    get cursorPositionContextStyle(): ConfigValue<'off' | 'inline' | 'lineAndColumn' | 'both'>
-    { return {settingID: 'cursorPositionContextStyle', value: CONFIG.cursorPositionContextStyle}; }
+    get cursorPositionContextStyle(): ConfigValue<'off' | 'inline' | 'lineAndColumn' | 'both'> { return { settingID: 'cursorPositionContextStyle', value: CONFIG.cursorPositionContextStyle }; }
 
-    get lineNumberContextFormat(): ConfigValue<string>
-    { return {settingID: 'lineNumberContextFormat', value: CONFIG.lineNumberContextFormat}; }
+    get lineNumberContextFormat(): ConfigValue<string> { return { settingID: 'lineNumberContextFormat', value: CONFIG.lineNumberContextFormat }; }
 
-    get websocketUrl(): ConfigValue<string>
-    { return {settingID: 'connection.websocketUrl', value: CONNECTION.websocketUrl}; }
+    get websocketUrl(): ConfigValue<string> { return { settingID: 'connection.websocketUrl', value: CONNECTION.websocketUrl }; }
 
-    get gameName(): ConfigValue<string>
-    { return {settingID: 'connection.gameName', value: CONNECTION.gameName}; }
+    get gameName(): ConfigValue<string> { return { settingID: 'connection.gameName', value: CONNECTION.gameName }; }
 
-    get userName(): ConfigValue<string>
-    { return {settingID: 'connection.userName', value: CONNECTION.userName}; }
+    get userName(): ConfigValue<string> { return { settingID: 'connection.userName', value: CONNECTION.userName }; }
 
-    get nameOfAPI(): ConfigValue<string>
-    { return {settingID: 'connection.nameOfAPI', value: CONNECTION.nameOfAPI}; }
+    get nameOfAPI(): ConfigValue<string> { return { settingID: 'connection.nameOfAPI', value: CONNECTION.nameOfAPI }; }
 }
 
 export const api: NeuroPilotAPI = {
@@ -55,21 +45,20 @@ export const api: NeuroPilotAPI = {
         failure: actionHandlerFailure,
         retry: actionHandlerRetry,
     },
-    getActions(action?: string | string[]) {
-        if (typeof action === 'string') {
-            const foundAction = getAction(action);
-            if (!foundAction) return foundAction;
-            const actionReturn = {...foundAction, source: findByToken(foundAction.sourceToken)?.name };
-            delete actionReturn?.sourceToken;
-            return actionReturn;
-        } else {
-            const actions = getActions(action);
-            return actions.map((a) => {
-                const newObject = {...a, source: findByToken(a.sourceToken)?.name };
-                delete newObject.sourceToken;
-                return newObject;
-            });
-        };
+    getAction(action: string) {
+        const foundAction = getAction(action);
+        if (!foundAction) return foundAction;
+        const actionReturn = { ...foundAction, source: findByToken(foundAction.sourceToken)?.name };
+        delete actionReturn?.sourceToken;
+        return actionReturn;
+    },
+    getActions(action: string[]) {
+        const actions = getActions(action);
+        return actions.map((a) => {
+            const newObject = { ...a, source: findByToken(a.sourceToken)?.name };
+            delete newObject.sourceToken;
+            return newObject;
+        });
     },
     getPositionContext,
     formatContext,